@@ -1,8 +1,7 @@
 package me.ehp246.aufjms.api.dispatch;
 
-import javax.jms.Connection;
-import javax.jms.JMSProducer;
-
+import jakarta.jms.Connection;
+import jakarta.jms.JMSProducer;
 import me.ehp246.aufjms.api.jms.JmsMsg;
 
 /**
@@ -10,7 +9,7 @@ import me.ehp246.aufjms.api.jms.JmsMsg;
  * {@link Connection}.
  * <p>
  * The interface only is concerned with sending the message.
- * 
+ *
  * @author Lei Yang
  *
  */
