@@ -1,10 +1,10 @@
 package me.ehp246.aufjms.api.jms;
 
-import javax.jms.JMSException;
-import javax.jms.JMSRuntimeException;
+import jakarta.jms.JMSException;
+import jakarta.jms.JMSRuntimeException;
 
 /**
- * 
+ *
  * @author Lei Yang
  * @since 1.0
  */
@@ -12,10 +12,10 @@ import javax.jms.JMSRuntimeException;
 public interface JMSSupplier<V> {
     V get() throws JMSException;
 
-    static <V> V invoke(JMSSupplier<V> callable) {
+    static <V> V invoke(final JMSSupplier<V> callable) {
         try {
             return callable.get();
-        } catch (JMSException e) {
+        } catch (final JMSException e) {
             throw new JMSRuntimeException(e.getMessage(), e.getErrorCode(), e);
         }
     }
