@@ -1,14 +1,13 @@
 package me.ehp246.aufjms.core.endpoint;
 
-import javax.jms.Destination;
-import javax.jms.JMSException;
-import javax.jms.JMSRuntimeException;
-import javax.jms.Queue;
-import javax.jms.Topic;
-
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import jakarta.jms.Destination;
+import jakarta.jms.JMSException;
+import jakarta.jms.JMSRuntimeException;
+import jakarta.jms.Queue;
+import jakarta.jms.Topic;
 import me.ehp246.aufjms.api.dispatch.JmsDispatch;
 import me.ehp246.aufjms.api.dispatch.JmsDispatchFn;
 import me.ehp246.aufjms.api.endpoint.InvocationListener;
@@ -20,12 +19,12 @@ final class ReplyInvoked implements InvocationListener.OnCompleted {
 
     private final JmsDispatchFn dispatchFn;
 
-    ReplyInvoked(JmsDispatchFn dispatchFn) {
+    ReplyInvoked(final JmsDispatchFn dispatchFn) {
         this.dispatchFn = dispatchFn;
     }
 
     @Override
-    public void onCompleted(Completed completed) {
+    public void onCompleted(final Completed completed) {
         final var msg = completed.bound().msg();
         final var replyTo = msg.replyTo();
         if (replyTo == null) {
@@ -42,7 +41,7 @@ final class ReplyInvoked implements InvocationListener.OnCompleted {
         try {
             return replyTo instanceof Queue ? At.toQueue(((Queue) replyTo).getQueueName())
                     : At.toTopic(((Topic) replyTo).getTopicName());
-        } catch (JMSException e) {
+        } catch (final JMSException e) {
             throw new JMSRuntimeException(e.getMessage(), e.getErrorCode(), e);
         }
     }
