@@ -2,8 +2,6 @@ package me.ehp246.aufjms.integration.dispatch.type;
 
 import java.util.Map;
 
-import javax.jms.JMSException;
-
 import org.jgroups.util.UUID;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -12,6 +10,7 @@ import org.junit.jupiter.api.Timeout;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import jakarta.jms.JMSException;
 import me.ehp246.aufjms.integration.dispatch.type.AppConfig.OfPropertyCase01;
 import me.ehp246.aufjms.util.EmbeddedArtemisConfig;
 import me.ehp246.aufjms.util.TestQueueListener;
