@@ -1,7 +1,5 @@
 package me.ehp246.aufjms.integration.dispatch.type;
 
-import javax.jms.JMSException;
-
 import org.jgroups.util.UUID;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,6 +8,7 @@ import org.junit.jupiter.api.Timeout;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import jakarta.jms.JMSException;
 import me.ehp246.aufjms.integration.dispatch.type.AppConfig.OfTypeCase01;
 import me.ehp246.aufjms.util.EmbeddedArtemisConfig;
 import me.ehp246.aufjms.util.TestQueueListener;
