@@ -5,10 +5,6 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
 
-import javax.jms.JMSException;
-import javax.jms.Message;
-import javax.jms.Session;
-
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.ListableBeanFactory;
@@ -18,6 +14,9 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.jms.core.MessageCreator;
 
+import jakarta.jms.JMSException;
+import jakarta.jms.Message;
+import jakarta.jms.Session;
 import me.ehp246.aufjms.api.endpoint.InboundEndpoint;
 import me.ehp246.aufjms.util.TestQueueListener;
 
@@ -44,7 +43,7 @@ class PropertyTest {
         jmsTemplate.send(TestQueueListener.DESTINATION_NAME, new MessageCreator() {
 
             @Override
-            public Message createMessage(Session session) throws JMSException {
+            public Message createMessage(final Session session) throws JMSException {
                 final var msg = session.createTextMessage();
                 msg.setBooleanProperty("b1", false);
                 msg.setJMSCorrelationID(UUID.randomUUID().toString());
@@ -59,7 +58,7 @@ class PropertyTest {
     void failureInterceptor_01() {
         Assertions.assertEquals(null,
                 beanFactory.getBean("InboundEndpoint-0", InboundEndpoint.class).invocationListener());
-        
+
         Assertions.assertEquals(appConfig.inteceptor,
                 beanFactory.getBean("InboundEndpoint-1", InboundEndpoint.class).invocationListener());
 
