@@ -1,13 +1,12 @@
 package me.ehp246.aufjms.util;
 
-import javax.jms.ConnectionFactory;
-
 import org.apache.qpid.jms.JmsConnectionFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.jms.connection.SingleConnectionFactory;
 
+import jakarta.jms.ConnectionFactory;
 import me.ehp246.aufjms.api.spi.JacksonConfig;
 
 /**
