@@ -2,14 +2,13 @@ package me.ehp246.aufjms.util;
 
 import java.util.concurrent.CompletableFuture;
 
-import javax.jms.ConnectionFactory;
-import javax.jms.Message;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.jms.config.DefaultJmsListenerContainerFactory;
 import org.springframework.jms.config.JmsListenerContainerFactory;
 
+import jakarta.jms.ConnectionFactory;
+import jakarta.jms.Message;
 import me.ehp246.aufjms.core.util.OneUtil;
 
 /**
@@ -23,7 +22,7 @@ public class TestTopicListener {
     private CompletableFuture<Message> received = new CompletableFuture<>();
 
     @JmsListener(destination = TestTopicListener.DESTINATION_NAME, subscription = SUBSCRIPTION_NAME)
-    public void receive(Message message) {
+    public void receive(final Message message) {
         received.complete(message);
     }
 
@@ -37,7 +36,7 @@ public class TestTopicListener {
 
     @Bean
     public JmsListenerContainerFactory<?> jmsListenerContainerFactory(final ConnectionFactory connectionFactory) {
-        DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
+        final DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
         factory.setConnectionFactory(connectionFactory);
         return factory;
     }
